+++ src/pikepdf/objects.py
@@ -256,7 +256,6 @@ class Stream(Object, metaclass=_ObjectMeta):
             kwargs: Keyword arguments that will define the stream dictionary. Do not set
                 /Length here as pikepdf will manage this value. Set /Filter
                 if the data is already encoded in some format.
-            obj: Deprecated alias for *data*.
         Returns:
             pikepdf.Object
 
@@ -279,15 +278,10 @@ class Stream(Object, metaclass=_ObjectMeta):
 
         .. versionchanged:: 2.2
             Support creation of ``pikepdf.Stream`` from existing dictionary.
-        """
 
-        # Support __new__(...obj=bytes) which should have been data=bytes,
-        # drop in pikepdf 3
-        if 'obj' in kwargs:  # pragma: no cover
-            warn("Deprecated parameter 'obj', use 'data' instead", DeprecationWarning)
-            if data is None:
-                data = kwargs['obj']
-            del kwargs['obj']
+        .. versionchanged:: 3.0
+            Deprecated ``obj`` argument was removed; use ``data``.
+        """
 
         if data is None:
             raise TypeError("Must make Stream from binary data")